Upadted cypress, added commands to execute e2e tests in chrome, firefox and webkit(safari), fixed some e2e tests
