+++ cypress/e2e/1-sign-up/signup-password-requirements.cy.ts
