+++ cypress/e2e/2-upload-items/drag-and-drop.cy.ts
@@ -1,6 +1,6 @@
 /* eslint-disable cypress/unsafe-to-chain-command */
 import { join } from 'path';
-import { FILE_ITEM_SELECTOR, MENU_ITEM_SELECTOR } from '../constans';
+import { FILE_ITEM_SELECTOR, MENU_ITEM_SELECTOR } from '../../constans';
 
 describe('Drag and drop', () => {
   const filename = 'example.txt';
