+++ cypress/e2e/6-preferences/account-panel.cy.ts
