+++ cypress/fixtures/example.txt
