+++ cypress/fixtures/photos/common/mock-info-1.json
