+++ cypress/fixtures/photos/common/mock-info-2.json
