+++ cypress/fixtures/photos/common/mock-photo-1.blob
