+++ cypress/fixtures/photos/common/mock-photo-2.blob
