+++ cypress/fixtures/photos/common/mock-preview-info-1.json
