+++ cypress/fixtures/photos/common/mock-preview-info-2.json
