+++ cypress/fixtures/photos/common/mock-preview-photo-1.blob
