+++ cypress/fixtures/photos/common/mock-preview-photo-2.blob
