+++ cypress/fixtures/photos/common/mock-results-1.json
