+++ cypress/fixtures/photos/common/mock-results-multiple.json
