+++ cypress/utils/photos.mock.ts
