+++ package.json
@@ -88,8 +88,11 @@
     "build": "craco build",
     "vercel:install": "yarn run add:npmrc && yarn install",
     "test:unit": "jest src/ && jest test/unit",
-    "test:e2e": "node ./cypress-transform-dotenv.js && cypress run --project test/e2e",
-    "cypress:open": "node ./cypress-transform-dotenv.js && cypress open --project test/e2e",
+    "test:e2e": "node ./cypress-transform-dotenv.js && cypress run",
+    "test:chrome": "node ./cypress-transform-dotenv.js && cypress run --browser chrome",
+    "test:firefox": "node ./cypress-transform-dotenv.js && cypress run --browser firefox",
+    "test:safari": "node ./cypress-transform-dotenv.js && cypress run --browser webkit",
+    "cypress:open": "node ./cypress-transform-dotenv.js && cypress open",
     "test": "npm run test:unit && npm run test:e2e",
     "eject": "react-scripts eject",
     "format": "prettier \"{src,test}/**/*.{ts,tsx}\" --write",
@@ -144,12 +147,13 @@
     "cross-env": "^7.0.3",
     "crypto": "npm:crypto-browserify",
     "customize-cra": "^1.0.0",
-    "cypress": "^9.7.0",
+    "cypress": "^13.9.0",
     "cypress-file-upload": "^5.0.8",
     "dotenv": "^10.0.0",
     "eslint": "^7.32.0",
-    "eslint-plugin-cypress": "^2.15.1",
+    "eslint-plugin-cypress": "^3.2.0",
     "husky": "^7.0.2",
+    "playwright-webkit": "1.34",
     "postcss": "^8.4.31",
     "prettier": "^2.5.0",
     "prettier-plugin-tailwindcss": "^0.1.8",
