+++ src/app/backups/components/BackupsAsFoldersList/BackupsAsFoldersList.tsx
@@ -1,23 +1,23 @@
+import { DriveFolderData } from '@internxt/sdk/dist/drive/storage/types';
+import folderEmptyImage from 'assets/icons/light/folder-open.svg';
+import _ from 'lodash';
 import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import _ from 'lodash';
-import { DriveFolderData } from '@internxt/sdk/dist/drive/storage/types';
 import { SdkFactory } from '../../../core/factory/sdk';
+import dateService from '../../../core/services/date.service';
+import { contextMenuSelectedBackupItems } from '../../../drive/components/DriveExplorer/DriveExplorerList/DriveItemContextMenu';
+import DriveListItemSkeleton from '../../../drive/components/DriveListItemSkeleton/DriveListItemSkeleton';
+import { deleteFile } from '../../../drive/services/file.service';
+import { deleteBackupDeviceAsFolder } from '../../../drive/services/folder.service';
+import iconService from '../../../drive/services/icon.service';
+import sizeService from '../../../drive/services/size.service';
+import { DriveItemData, DriveFolderData as DriveWebFolderData } from '../../../drive/types';
+import { useTranslationContext } from '../../../i18n/provider/TranslationProvider';
 import Empty from '../../../shared/components/Empty/Empty';
-import { DriveFolderData as DriveWebFolderData, DriveItemData } from '../../../drive/types';
+import List from '../../../shared/components/List';
 import { deleteItemsThunk } from '../../../store/slices/storage/storage.thunks/deleteItemsThunk';
-import folderEmptyImage from 'assets/icons/light/folder-open.svg';
 import { downloadItemsThunk } from '../../../store/slices/storage/storage.thunks/downloadItemsThunk';
 import { uiActions } from '../../../store/slices/ui';
-import DriveListItemSkeleton from '../../../drive/components/DriveListItemSkeleton/DriveListItemSkeleton';
-import { deleteBackupDeviceAsFolder } from '../../../drive/services/folder.service';
-import { deleteFile } from '../../../drive/services/file.service';
-import List from '../../../shared/components/List';
-import { contextMenuSelectedBackupItems } from '../../../drive/components/DriveExplorer/DriveExplorerList/DriveItemContextMenu';
-import { useTranslationContext } from '../../../i18n/provider/TranslationProvider';
-import iconService from '../../../drive/services/icon.service';
-import dateService from '../../../core/services/date.service';
-import sizeService from '../../../drive/services/size.service';
 
 export default function BackupsAsFoldersList({
   className = '',
@@ -132,7 +132,10 @@ export default function BackupsAsFoldersList({
                 const Icon = iconService.getItemIcon(item.isFolder, item.type);
 
                 return (
-                  <div className="flex min-w-activity grow items-center justify-start pr-3">
+                  <div
+                    className="flex min-w-activity grow items-center justify-start pr-3"
+                    data-test="backup-list-folder"
+                  >
                     <div className="mr-3 h-8 w-8">
                       <Icon className="h-8 w-8" />
                     </div>
