--- test/e2e/cypress.json
@@ -1,32 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "chromeWebSecurity": false,
-  "defaultCommandTimeout": 60000,
-  "pluginsFile": "cypress/plugins/index.ts",
-  "modifyObstructiveCode": false,
-  "video": false,
-  "testFiles": [
-    "signup.spec.ts",
-    "signup-password-requirements.spec.ts",
-    "signup-onboarding.spec.ts",
-    "unauthorized.spec.ts",
-    "upload-file.spec.ts",
-    "drive-pagination.spec.ts",
-    "drive-icon-list.spec.ts",
-    "recents-panel.spec.ts",
-    "account-panel.spec.ts",
-    "backups.spec.ts",
-    "download-file.spec.ts",
-    "share-links.spec.ts",
-    "rename-file.spec.ts",
-    "download-folder.spec.ts",
-    "remove-folder.spec.ts",
-    "delete-file.spec.ts",
-    "drag-and-drop.spec.ts",
-    "photos.spec.ts",
-    "change-password.spec.ts"
-  ],
-  "viewportWidth": 1500,
-  "viewportHeight": 660,
-  "experimentalSessionAndOrigin": true
-}
