--- test/e2e/cypress/.eslintrc.json
@@ -1,8 +1,0 @@
-{
-  "plugins": [
-    "cypress"
-  ],
-  "extends": [
-    "plugin:cypress/recommended"
-  ]
-}
