--- test/e2e/cypress/fixtures/test-user.json
@@ -1,1 +1,0 @@
-{"username":"513ecde086e0d066dbc245fe7d226a5fc1669e10cad003c16d0457f8dd9aa38e@inxt.com","password":"Pw48ca9137f"}
