--- test/e2e/cypress/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "lib": ["es5", "dom"],
-    "types": ["cypress", "node", "./support"]
-  },
-  "include": ["**/*.ts"]
-}
